Crud
Implementing a simple CRUD

+++ BusinessCard/BusinessCardCore/BaseEntity.cs
@@ -8,7 +8,7 @@ namespace BusinessCardCore
 {
     public abstract class BaseEntity
     {
-        public Int64 ID { get; set; }
+        public int ID { get; set; }
         public DateTime AddedDate { get; set; }
         public DateTime ModifiedDate { get; set; }
         public string AddedUser { get; set; }
